♻️ REFACTOR: 파일명 index로 바꿔서 경로 단축 시키기
