+++ src/pages/Main/index.js
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useState } from "react";
 import CardList from "@components/CardList/CardList";
 import ProfileImage from "@components/ProfileImage/ProfileImage";
 import "./Main.css";
-import { SearchBar } from "@components/Input/SearchBarStyle";
+import { SearchBar } from "@components/SearchBar";
 import { NO_LINK_FOUND } from "@/constants";
 import { getFolderData } from "@/api/api";
